+++ tools/visual/source/perlin_noise/perlin_noise.hpp
@@ -22,6 +22,7 @@
 #include <vector>
 #include <string>
 #include <memory>
+#include <fstream>
 
 #include <boost/filesystem.hpp>
 
@@ -36,14 +37,14 @@ namespace itl
     public:
         explicit PerlinNoise(const std::shared_ptr<Logger>& log,
                              const std::string& path_to_data);
-        const std::vector<float>& get_random_noise() const;
+        const std::vector<double>& get_random_noise() const;
     private:
         void generate_noise_2d(const std::string& dir);
-        float generate_point_noise(float x, float y, float z, const std::vector<int>& permutation) const;
-        void save_to_file(const std::vector<float>& noise, const std::string& dir) const;
+        double generate_point_noise(double x, double y, double z, const std::vector<int>& permutation) const;
+        void save_to_file(const std::vector<double>& noise, const std::string& dir) const;
         void read_from_file(const std::string& dir);
 
-        std::vector<std::vector<float>> noises;
+        std::vector<std::vector<double>> noises;
         std::shared_ptr<Logger> logger;
     };
 }
